requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
moviepy==1.0.3
openai-whisper==20231117
watchdog==3.0.0
loguru==0.7.2
ratelimit==2.2.1
opencv-python==4.9.0.80
numpy==1.26.4
openai==1.12.0
scikit-learn==1.4.0
Pillow==10.2.0
psutil==5.9.8
PyYAML==6.0.1
cryptography==42.0.2
tqdm==4.66.2
fastapi==0.109.2
uvicorn==0.27.1
prometheus-client==0.19.0
gputil==1.4.0
boto3==1.34.34
requests==2.31.0
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
tabulate==0.9.0
weasyprint==60.2
torch==2.2.1
torchvision==0.17.1
transformers==4.38.2
librosa==0.10.1
scipy==1.12.0 